+++ packages/http/src/contract.ts
@@ -1,9 +1,9 @@
 import type { Property } from '@sonata-api/types'
 
-export type RouteContract = Property | [
-  Property | null,
-  Property | Property[] | null
-]
+export type RouteContract = Property | {
+  request: Property | null
+  response: Property | Property[] | null
+}
 
 export const defineContract = <const TRouteContract extends RouteContract>(contract: TRouteContract) => {
   return contract