+++ packages/server/src/init.ts
@@ -1,33 +1,33 @@
 import type { Context, GenericRequest, ApiConfig, DecodedToken } from '@sonata-api/types'
-import { right, left, isLeft, unwrapEither } from '@sonata-api/common'
+import { right, left, isLeft, unwrapEither, unsafe } from '@sonata-api/common'
 import { defineServerOptions, cors, wrapRouteExecution } from '@sonata-api/http'
 import { registerServer } from '@sonata-api/node-http'
 
-import { createContext, decodeToken, ObjectId } from '@sonata-api/api'
+import { createContext, decodeToken, traverseDocument } from '@sonata-api/api'
 import { getDatabase } from '@sonata-api/api'
 import { DEFAULT_API_CONFIG } from './constants'
 import { warmup } from './warmup'
 import { registerRoutes } from './routes'
 
-export const getDecodedToken = async (request: GenericRequest) => {
-  try {
-    const decodedToken: DecodedToken = request.headers.authorization
-      ? {
-        authenticated: true,
-        ...await decodeToken(request.headers.authorization.split('Bearer ').pop() || '')
+export const getDecodedToken = async (request: GenericRequest, context: Context) => {
+  if( !request.headers.authorization ) {
+    return right(<DecodedToken>{
+      authenticated: false,
+      user: {
+        _id: null
       }
-      : {
-        authenticated: false,
-        user: {
-          _id: null
-        }
-      }
+    })
+  }
 
-      if( decodedToken.authenticated ) {
-        decodedToken.user._id = new ObjectId(decodedToken.user._id)
-      }
+  try {
+    const decodedToken: DecodedToken = await decodeToken(request.headers.authorization.split('Bearer ').pop() || '')
+    decodedToken.authenticated = true
+    decodedToken.user = unsafe(await traverseDocument(decodedToken.user, context.collections.user.description, {
+      autoCast: true
+    }))
 
     return right(decodedToken)
+
   } catch( err ) {
     if( process.env.NODE_ENV === 'development' ) {
       console.trace(err)
@@ -67,7 +67,7 @@ export const dryInit = async (
     }
 
     await wrapRouteExecution(response, async () => {
-      const tokenEither = await getDecodedToken(request)
+      const tokenEither = await getDecodedToken(request, parentContext)
       if( isLeft(tokenEither) ) {
         return tokenEither
       }